Explicitly reset timer.
This fixes the component spinning too long when the promise is resolved before
the spinner is created. This typically happens in tests or if the startDelay is
too long. The promise getting resolved triggers the setEnabled function before
createSpinner is called, and the component spins until the timeout is reached.